AMLS-5181 | Refactor
